+++ packages/core/src/components/Toggle/Toggle.styled.tsx
@@ -1,18 +1,32 @@
 import { defaultTheme } from '@medly-components/theme';
 import { WithThemeProp } from '@medly-components/utils';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import FieldWithLabel from '../FieldWithLabel';
 import { ToggleProps, ToggleWrapperProps } from './types';
 
 const getHeight = ({ theme, size }: ToggleProps & WithThemeProp) =>
     size ? theme.toggle.sizes[size] : theme.toggle.sizes[theme?.toggle.defaultSize];
+
 const getWidth = (props: ToggleProps & WithThemeProp) => `calc(${getHeight(props)} * 2)`;
+
+const IconStyle = css`
+    color: ${({ theme }) => theme.toggle.iconColor};
+    transition: 0.2s ease-in-out;
+    position: absolute;
+    svg {
+        height: ${props => `calc(${getHeight(props)} - 1rem) `};
+        path {
+            fill: ${({ theme }) => theme.toggle.iconColor};
+        }
+    }
+`;
 
 export const Wrapper = styled(FieldWithLabel.Field)<ToggleWrapperProps>`
     position: relative;
     width: ${getWidth};
     height: ${getHeight};
-    border-radius: 15px;
+    border-radius: 1.5rem;
+    background-color: ${({ theme }) => theme.toggle.bgColor};
 
     &:focus-within {
         box-shadow: 0 0 0 1pt ${({ theme }) => theme.toggle.outlineColor};
@@ -21,30 +35,45 @@ export const Wrapper = styled(FieldWithLabel.Field)<ToggleWrapperProps>`
     & > * {
         width: 100%;
         height: 100%;
-        border-radius: 15px;
+        border-radius: 1.5rem;
     }
 `;
+
 Wrapper.defaultProps = {
     theme: defaultTheme
 };
 
+export const Switch = styled('div')<ToggleWrapperProps>`
+    position: relative;
+    width: 100%;
+    overflow: hidden;
+    display: flex;
+    align-items: flex-end;
+`;
+
+export const OffIcon = styled('div')<ToggleWrapperProps>`
+    ${IconStyle}
+    right: 0;
+`;
+
+export const OnIcon = styled('div')<ToggleWrapperProps>`
+    ${IconStyle}
+    opacity: 0;
+    left: ${props => `calc((-${getHeight(props)}) + 0.3rem) `};
+`;
+
 export const Circle = styled('div')<ToggleWrapperProps>`
     z-index: 1;
     position: absolute;
-    background-color: ${({ theme }) => theme.toggle.bgColor};
-
-    &::after {
-        content: '';
-        margin: 3px;
-        display: block;
-        border-radius: 50%;
-        background: ${({ theme }) => theme.colors.white};
-        box-shadow: 1px 3px 3px 1px rgba(0, 0, 0, 0.2);
-        transition: 0.2s;
-        width: ${props => `calc(${getHeight(props)} - 6px) `};
-        height: ${props => `calc(${getHeight(props)} - 6px) `};
-    }
+    height: ${props => `calc(${getHeight(props)} - 0.6rem) `};
+    width: ${props => `calc(${getHeight(props)} - 0.6rem) `};
+    margin: 0.3rem;
+    border-radius: 50%;
+    background: ${({ theme }) => theme.colors.white};
+    box-shadow: 0.1rem 0.3rem 0.3rem 0.1rem rgba(0, 0, 0, 0.2);
+    transition: 0.2s;
 `;
+
 Circle.defaultProps = {
     theme: defaultTheme
 };
@@ -62,15 +91,28 @@ export const Checkbox = styled('input').attrs({ type: 'checkbox' })<ToggleProps>
 
     &:disabled {
         cursor: not-allowed;
-        & + ${Circle} {
+        & + ${Switch} {
             background: ${({ theme }) => theme.toggle.disabledBgColor};
         }
     }
 
-    &:checked + ${Circle} {
+    &:checked + ${Switch} {
         background: ${({ theme }) => theme.toggle.checkedBgColor};
-        &::after {
-            margin-left: ${props => `calc(${getWidth(props)} - (${getHeight(props)}) + 3px) `};
+
+        ${Circle} {
+            margin-left: ${props => `calc((${getHeight(props)}) + 0.3rem) `};
+        }
+
+        ${OffIcon} {
+            color: ${({ theme }) => theme.toggle.iconColor};
+            opacity: 0;
+            right: ${props => `calc((-${getHeight(props)}) + 0.3rem) `};
+        }
+
+        ${OnIcon} {
+            opacity: 1;
+            position: absolute;
+            left: 0;
         }
     }
 `;