+++ packages/core/src/components/Toggle/Toggle.test.tsx
@@ -1,3 +1,4 @@
+import { LockOpenIcon, LockOutlineIcon } from '@medly-components/icons';
 import { fireEvent, render, screen } from '@test-utils';
 import { Toggle } from './Toggle';
 
@@ -19,5 +20,23 @@ describe('Checkbox component', () => {
         render(<Toggle onChange={mockOnChange} disabled name="gender" label="Female" size="M" labelPosition="right" />);
         fireEvent.click(screen.getByRole('checkbox'));
         expect(mockOnChange).toBeCalled();
+    });
+
+    it('should show the toggle button icons when icons are passed', () => {
+        const mockOnChange = jest.fn();
+        render(
+            <Toggle
+                onChange={mockOnChange}
+                disabled
+                name="gender"
+                label="Female"
+                size="M"
+                labelPosition="right"
+                iconLeft={<LockOpenIcon data-testid="lock-open-icon" />}
+                iconRight={<LockOutlineIcon data-testid="lock-close-icon" />}
+            />
+        );
+        expect(screen.getByTestId('lock-open-icon')).toBeInTheDocument();
+        expect(screen.getByTestId('lock-close-icon')).toBeInTheDocument();
     });
 });