+++ packages/markdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.23](https://github.com/medly/medly-components/compare/@medly-components/markdown@0.4.22...@medly-components/markdown@0.4.23) (2022-06-28)
+
+**Note:** Version bump only for package @medly-components/markdown
+
+
+
+
+
 ## [0.4.22](https://github.com/medly/medly-components/compare/@medly-components/markdown@0.4.21...@medly-components/markdown@0.4.22) (2022-06-02)
 
 **Note:** Version bump only for package @medly-components/markdown