+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/medly/medly-components/compare/@medly-components/utils@2.1.2...@medly-components/utils@2.1.3) (2022-06-28)
+
+**Note:** Version bump only for package @medly-components/utils
+
+
+
+
+
 ## [2.1.2](https://github.com/medly/medly-components/compare/@medly-components/utils@2.1.1...@medly-components/utils@2.1.2) (2022-06-02)
 
 **Note:** Version bump only for package @medly-components/utils